InfoPanelCommon を削除、PartsEditProps を削除、PartsEditComponent を追加
