--- packages/maker/src/info/InfoPanelCommon.module.scss
@@ -1,15 +1,0 @@
-.row {
-    display: flex;
-}
-
-.rowTitle {
-    flex: 0 0 auto;
-}
-
-.rowContent {
-    flex: 1 1 auto;
-}
-
-.grid {
-    display: grid;
-}
